requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
dnspython==2.8.0
fastapi==0.119.0
graphql-core==3.2.6
h11==0.16.0
idna==3.11
Jinja2==3.1.6
lia-web==0.2.3
MarkupSafe==3.0.3
motor==3.7.1
packaging==25.0
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
strawberry-graphql==0.284.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
